backend/requirements.txt
# Core build dependencies
setuptools==65.5.1
wheel

# Project Dependencies
# Core Application Dependencies
pydantic==2.11.7
supabase
fastapi==0.116.1
uvicorn[standard]==0.29.0
mangum==0.19.0
aiohttp==3.12.15
typing-extensions==4.15.0
httpx==0.28.1
anyio==4.8.0
certifi==2024.7.4
httpcore==1.0.5
idna==3.7
sniffio==1.3.1
h11==0.14.0
starlette==0.46.0
exceptiongroup==1.1.2
click==8.1.7
aiohappyeyeballs==2.5.0
aiosignal==1.4.0
async-timeout==4.0.3
attrs==23.2.0
frozenlist==1.4.1  # Updated to a Python 3.12 compatible version
multidict==6.0.5
propcache==0.2.1
yarl==1.17.0
aiofiles==23.2.1
aiosqlite==0.17.0
supabase_auth==2.12.3
pyjwt==2.10.1
strenum<0.5.0,>=0.4.15
realtime==2.7.0
postgrest==1.1.1
deprecation==2.1.0
python-dotenv==1.0.1
uvloop==0.19.0
watchfiles==0.21.0
supabase-functions==0.10.1

# ML/NLP and Data Processing Dependencies
asyncpraw
boto3==1.40.17
websockets==13.0.0
numpy
scipy
scikit-learn
pandas
Cython
cymem
preshed
murmurhash==1.0.8
protobuf==3.19.6
absl-py==1.4.0
grpcio==1.74.0

# grpcio-status will be resolved by grpcio
tokenizers==0.19.1
sentencepiece==0.1.96
transformers==4.42.0
sentence-transformers==2.7.0
huggingface-hub==0.23.4
spacy==3.7.4
annotated-types==0.6.0
pydantic-core==2.33.2
typing-inspection==0.4.1
update_checker==0.18.0
botocore==1.40.17
jmespath==0.10.0
s3transfer==0.13.0
python-dateutil==2.8.2
urllib3==1.26.18
six==1.16.0
joblib==1.4.2
threadpoolctl==3.1.0
filelock==3.12.2
fsspec==2024.6.1
packaging==23.2
PyYAML==6.0.1

requests==2.32.3
tqdm==4.66.2
regex==2024.5.15
safetensors==0.4.3
torch
Pillow==9.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
thinc==8.2.5
wasabi==1.1.2
srsly==2.4.8
catalogue==2.0.8
weasel==0.3.0
typer==0.7.0
smart-open==6.4.0
Jinja2==3.1.2
langcodes==3.3.0
charset-normalizer==3.3.2
blis==0.7.11
confection==0.1.4
cloudpathlib==0.15.0
MarkupSafe==2.1.5
sympy==1.11.1
networkx==2.8.8

# Google API & Generative AI Specific Dependencies (Updated to new SDK)
google-genai
google-generativeai>=0.8.0
openai>=1.1.0
google-api-core==2.17.0
google-api-python-client==2.128.0
google-auth==2.29.0
proto-plus==1.23.0
googleapis-common-protos==1.56.4
cachetools==5.3.3
pyasn1-modules==0.3.0
httplib2==0.22.0
google-auth-httplib2==0.2.0
uritemplate==4.1.1
pyparsing==2.4.7
tenacity==8.2.3

# Other Utility Dependencies
brotli==1.1.0
psutil==7.0.0
rsa==4.1
zstandard==0.24.0
yake==0.4.8
nltk==3.9.1
pyasn1==0.5.1
tabulate==0.9.0
segtok==1.5.11
jellyfish==0.11.2
mpmath==1.2.1
h2==4.1.0
hyperframe==6.0.1
hpack==4.0.0
storage3==0.12.1
asyncprawcore==2.4.0
distro==1.7.0
jiter==0.4.0